+++ DDEyC_API/DataAccess/Models/DTOs/EmailRequestDTO.cs
@@ -10,11 +10,5 @@ namespace DDEyC_API.DataAccess.Models.DTOs
         public List<EmailAttachment> Attachments { get; set; } = new List<EmailAttachment>();
     }
 
-    public class EmailAttachment
-    {
-        public string FileName { get; set; }
-        public string ContentId { get; set; }
-        public byte[] Content { get; set; }
-        public string MimeType { get; set; }
-    }
+   
 }
